third_party/WebKit/LayoutTests/fast/dom/shadow/input-color-in-content.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/color_chooser/color_chooser.mojom.js"></script>
<script src="../../forms/color/mock-colorchooser.js"></script>
<div id="host" tabindex="1">
  <input id="target" type="color" value="#000000">
</div>
<script>

shadow = host.createShadowRoot();
shadow.innerHTML = "<content></content>";
jsTestIsAsync = true;

function ensureOpeningFor(count, then)
{
    shouldBeTrue("mockColorChooser.isChooserShown()");
    if (count)
        return window.setTimeout(function() { ensureOpeningFor(count -1, then); }, 0);
    then();
}

if (window.eventSender) {
   shouldBeFalse("mockColorChooser.isChooserShown()");
   eventSender.dragMode = false;
   eventSender.mouseMoveTo(target.offsetLeft, target.offsetTop + target.offsetHeight / 2);
   eventSender.mouseDown();
   eventSender.mouseUp();

   waitUntilChooserShown(function() {
      testRunner.setWindowIsKey(false); // This triggers a style recalc.
      ensureOpeningFor(10, finishJSTest); // Try several times to ensure nothing wrong happens asynchronously.
   });
}

</script>
</body>
</html>
