third_party/WebKit/LayoutTests/fast/dom/shadow/input-color-in-content-expected.txt

PASS mockColorChooser.isChooserShown() is false
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS successfullyParsed is true

TEST COMPLETE
